New IR pretty printer
Add an ir.pretty_print function that outputs a human-readable version of
the IR. This representation may be helpful for debugging because it is
significantly more concise than --dump=ir or --emit-c.

The main limitation is that we don't show any types. I couldn't think of a
clean way to show them so I just left them out.